backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
flask-marshmallow==1.1.0
marshmallow==3.20.1
marshmallow-sqlalchemy==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
gunicorn==21.2.0
supabase==2.0.3
